restructure functions into packages
