+++ src/utils/session_funcs.py
@@ -1,0 +1,106 @@
+# -*- coding: utf-8 -*-
+import os, random, requests
+from dotenv import load_dotenv
+# Load dotenv environment
+load_dotenv()
+NEWS_API_KEY = os.environ.get("NEWS_API_KEY")
+
+# ===============================================================
+# Header data for randomised referer and user-agent values
+# ===============================================================
+
+REFERERS = ["https://www.google.com/",
+            "https://bing.com/",
+            "https://search.yahoo.com/",
+            "https://www.baidu.com/",
+            "https://yandex.com/"]
+
+REFERER_PROBS = [0.88, 0.03, 0.03, 0.03, 0.03]
+
+USER_AGENTS = [
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/117.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/117.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/118.0",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/117.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (X11; Linux x86_64; rv:109.0) Gecko/20100101 Firefox/118.0",
+    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:109.0) Gecko/20100101 Firefox/118.0",
+    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/17.0 Safari/605.1.15",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/117.0",
+    "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:109.0) Gecko/20100101 Firefox/118.0",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36 OPR/102.0.0.0",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/117.0.0.0 Safari/537.36 Edg/117.0.2045.60",
+    "Mozilla/5.0 (Windows NT 10.0; rv:109.0) Gecko/20100101 Firefox/118.0",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36 Edg/118.0.2088.46",
+    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.6 Safari/605.1.15",
+    "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36",
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/117.0.0.0 Safari/537.36 Edg/117.0.2045.47"
+]
+
+USER_AGENT_PROBS = [
+    0.205, 0.14, 0.13, 0.105, 0.055,
+    0.055, 0.05, 0.045, 0.04, 0.03,
+    0.025, 0.02, 0.015, 0.015, 0.015,
+    0.0125, 0.0125, 0.012, 0.01, 0.008
+]
+
+# ===============================================================
+# API request functions
+# ===============================================================
+
+def weighted_random_selection(sample_space: list[str], probs: list[float]) -> str:
+    """
+    Assigns a weighted-random selection of header data
+    for API requests
+
+    Parameters
+    ----------
+    sample_space : list 
+        list of options to randomly select from
+    
+    probs : list 
+        list of probabilites per sample in sample_space (sum == 1)
+    
+    Returns
+    -------
+    weighted_random_selection : str
+        weighted random selection from sample_space
+    """
+    # k=1 to choose just one sample
+    weighted_random_selection = random.choices(sample_space, weights=probs, k=1)
+
+    return weighted_random_selection[0]
+
+
+def get_session(news_api=False) -> requests.Session:
+    """
+    Creates a new session for an API request
+
+    Parameters
+    ----------
+    news_api : bool
+        Flag to add API key for News API calls
+
+    Returns
+    -------
+    session : requests.Session
+        New session object for API requests
+    """
+    # Set session data
+    session = requests.Session()
+    # Get weighted random referer and user-agent values
+    referer = weighted_random_selection(REFERERS, REFERER_PROBS)
+    user_agent = weighted_random_selection(USER_AGENTS, USER_AGENT_PROBS)
+    # Assign values to session header
+    session.headers["User-Agent"] = user_agent
+    session.headers["Referer"] = referer
+    session.headers["Upgrade-Insecure-Requests"] = "1"
+
+    if news_api:
+      session.headers["X-Api-Key"] = NEWS_API_KEY
+    
+    return session
+
